BACKLOG-12428 : cleanup, refactoring and improvements
