+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTree/index.js
@@ -1,0 +1,3 @@
+import ContentTree from './ContentTree.container';
+
+export default ContentTree;
