--- src/javascript/JContent/ContentRoute/ContentLayout/ContentTrees/ContentTree/PickerViewMaterial/iconRenderer.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import {PageIcon, getIcon} from '@jahia/icons';
-
-function iconRenderer(entry) {
-    let Icon = getIcon(entry.node.primaryNodeType.name) || PageIcon;
-    return <Icon/>;
-}
-
-export default iconRenderer;
