--- src/javascript/JContent/ContentRoute/ContentLayout/ContentTrees/ContentTree/PickerViewMaterial/index.js
@@ -1,3 +1,0 @@
-import PickerViewMaterial from './PickerViewMaterial';
-
-export default PickerViewMaterial;
