--- src/javascript/JContent/ContentRoute/ContentLayout/ContentTrees/ContentTree/index.js
@@ -1,3 +1,0 @@
-import ContentTree from './ContentTree';
-
-export default ContentTree;
