--- src/javascript/JContent/ContentRoute/ContentLayout/ContentTrees/index.js
@@ -1,3 +1,0 @@
-import ContentTrees from './ContentTrees';
-
-export default ContentTrees;
