+++ src/javascript/JContent/JContent.accordion-items.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import {File, FolderSpecial, Collections, Setting} from '@jahia/moonstone/dist/icons';
-import ContentTrees from './ContentRoute/ContentLayout/ContentTrees/ContentTrees';
+import ContentTree from './ContentRoute/ContentLayout/ContentTree/ContentTree.container';
 import JContentConstants from './JContent.constants';
 
 export const jContentAccordionItems = registry => {
     const renderDefaultContentTrees = registry.add('accordionItem', 'renderDefaultContentTrees', {
         render: item => (
-            <ContentTrees contentTreeConfigs={item.configs}/>
+            <ContentTree contentTreeConfig={item.configs}/>
         )
     });
 
@@ -15,14 +15,14 @@ export const jContentAccordionItems = registry => {
         icon: <File/>,
         label: 'label.contentManager.navigation.pages',
         defaultUrl: siteKey => '/sites/' + siteKey,
-        configs: [{
+        configs: {
             rootPath: '',
             selectableTypes: ['jnt:page', 'jnt:virtualsite'],
             type: 'pages',
             openableTypes: ['jnt:page', 'jnt:virtualsite', 'jnt:navMenuText'],
             rootLabel: 'jcontent:label.contentManager.browsePages',
             key: 'browse-tree-pages'
-        }]
+        }
     });
 
     registry.add('accordionItem', JContentConstants.mode.CONTENT_FOLDERS, renderDefaultContentTrees, {
@@ -30,14 +30,14 @@ export const jContentAccordionItems = registry => {
         icon: <FolderSpecial/>,
         label: 'label.contentManager.navigation.contentFolders',
         defaultUrl: siteKey => '/sites/' + siteKey + '/contents',
-        configs: [{
+        configs: {
             rootPath: '/contents',
             selectableTypes: ['jmix:cmContentTreeDisplayable', 'jmix:visibleInContentTree', 'jnt:contentFolder'],
             type: 'contents',
             openableTypes: ['jmix:cmContentTreeDisplayable', 'jmix:visibleInContentTree', 'jnt:contentFolder'],
             rootLabel: 'jcontent:label.contentManager.browseFolders',
             key: 'browse-tree-content'
-        }]
+        }
     });
 
     registry.add('accordionItem', JContentConstants.mode.MEDIA, renderDefaultContentTrees, {
@@ -45,14 +45,14 @@ export const jContentAccordionItems = registry => {
         icon: <Collections/>,
         label: 'label.contentManager.navigation.media',
         defaultUrl: siteKey => '/sites/' + siteKey + '/files',
-        configs: [{
+        configs: {
             rootPath: '/files',
             selectableTypes: ['jnt:folder'],
             type: 'files',
             openableTypes: ['jnt:folder'],
             rootLabel: 'jcontent:label.contentManager.browseFiles',
             key: 'browse-tree-files'
-        }]
+        }
     });
 
     registry.add('accordionItem', JContentConstants.mode.APPS, {
